requirements.txt
alembic==1.0.9
Babel==2.6.0
blinker==1.4
Click==7.0
elasticsearch==7.0.0
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Babel==0.12.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.1
Werkzeug==0.15.2
WTForms==2.2.1
